% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_par_24_9_24h.R
\docType{data}
\name{best_par_24.9_24h}
\alias{best_par_24.9_24h}
\title{Set of parameters for a daily simulation at Tingvatn}
\format{A txt file}
\source{
\href{http://www.met.no}{met.no}
}
\description{
Set of parameters for a daily simulation at Tingvatn
}
\references{
the Norwegian Meteorological Institute (met.no)
}
\keyword{datasets}

