% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{bind_result}
\alias{bind_result}
\title{Binding extracted results in simulation}
\usage{
bind_result(result)
}
\arguments{
\item{result}{a list generated by pdd simulation function}
}
\value{
a binded list of extracted results
}
\description{
Function to bind all results extracted from pdd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
