% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{both_rates_vary}
\alias{both_rates_vary}
\title{Evaluate whether both speciation rates vary in specified DD model}
\usage{
both_rates_vary(ddmodel)
}
\arguments{
\item{ddmodel}{a character or integer specifying a DD model, as described in
dd_loglik() and dd_ML() documentation}
}
\value{
TRUE if both speciation and extinction rates vary with N, FALSE otherwise
}
\description{
Evaluate whether both speciation rates vary in specified DD model
}
\author{
Theo Pannetier
}
