% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{conv}
\alias{conv}
\title{Function to do convolution of two vectors}
\usage{
conv(x, y)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}
}
\value{
vector that is the convolution of x and y
}
\description{
Convolution of two vectors
}
\examples{

conv(1:10,1:10)

}
\references{
- Etienne, R.S. et al. 2012, Proc. Roy. Soc. B 279: 1300-1309,
doi: 10.1098/rspb.2011.1439 \cr - Etienne, R.S. & B. Haegeman 2012. Am. Nat.
180: E75-E89, doi: 10.1086/667574
}
\author{
Rampal S. Etienne
}
\keyword{models}
