% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_sim.R
\name{create_edd_pars}
\alias{create_edd_pars}
\title{Create a list of parameter sets for edd simulation}
\usage{
create_edd_pars(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical, decides whether to save the created list to file}

\item{...}{parameters that form a parameter space}
}
\value{
Returns a list containing the parameter sets
}
\description{
Function to conveniently create a list containing all possible
parameter combinations for edd simulation
}
\author{
Tianjian Qin
}
