% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{date}
\alias{date}
\title{ddd.date}
\usage{
date(timeResolution, fromPeriod, toPeriod, format, saveDate = NULL)
}
\arguments{
\item{timeResolution}{choice between "daily","hourly","minute","second"}

\item{fromPeriod}{"YYYYmmddHHMMSS"}

\item{toPeriod}{"YYYYmmddHHMMSS"}

\item{format}{conversion into a vector "YYMMDDHH" or a matrix "YY,MM,DD,HH"}

\item{saveDate}{"YYYYmmddHHMMSS"}
}
\description{
ddd.date
}
\examples{
res <- date(timeResolution="hourly",
            fromPeriod="20150215",
            toPeriod="20150218",
            format="YY,MM,DD,HH")
head(res)
}
\keyword{ddd}

