% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_main.R
\name{dd_main}
\alias{dd_main}
\title{Run the DD ruotine: simulation + MLE}
\usage{
dd_main(lambda, mu, K, age, cond = 1, seed, ddmodel = 1, output_file = NA)
}
\description{
Run the DD ruotine: simulation + MLE
}
\author{
Giovanni Laudanno
}
