% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dischargeNarsjo.R
\docType{data}
\name{dischargeNarsjo}
\alias{dischargeNarsjo}
\title{Discharge at Narsjo}
\format{A csv file}
\source{
\href{http://nve.no}{NVE}
}
\description{
Discharge at Narsjo measured by the Norwegian Water Resources and Energy Directorate
Hourly discharge are made available between August 2 2016 and September 29 2016.
No post-processing (qualicty control,...) has been done on the data.
}
\references{
the Norwegian Water Resources and Energy Directorate (NVE)
}
\keyword{datasets}

