% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.evapotranspiration.R
\name{do.evapotranspiration}
\alias{do.evapotranspiration}
\title{Computes the evapotranspiration}
\usage{
do.evapotranspiration(htemp, sca)
}
\arguments{
\item{htemp}{temperature for each level Zone}

\item{sca}{snow coverage for each level zone}
}
\description{
Computes the evapotranspiration.
}
\examples{
\dontrun{
do.evapotranspiration()
}
}
\keyword{ddd}

