% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.precipLZ.R
\name{do.precipLZ}
\alias{do.precipLZ}
\title{Precipitation for each elevation zone}
\usage{
do.precipLZ(v, modelPrecipLZ)
}
\arguments{
\item{v}{precipitation value (scalar)}

\item{modelPrecipLZ}{list(nbLevelZone,Plr,hfelt,midmetp)}
}
\description{
The function processes the precipitation for each elevation zone
}
\examples{
\dontrun{
do.precipLZ()
}
}
\keyword{ddd}

