% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.simTS.R
\name{do.simTS}
\alias{do.simTS}
\title{Simulation}
\usage{
do.simTS(timePeriod, q, precip, temp, scaob, snow, snowReservoir, soilMoisture,
  soilWater, UH, soilDischarge, ddistAll, groundwater, models, pathResults,
  saveDate)
}
\arguments{
\item{timePeriod}{data timeserie and its related information}

\item{q}{runoff timeserie}

\item{precip}{precipitation timeserie}

\item{temp}{temperature timeserie}

\item{scaob}{observed snow coverage timeserie}

\item{snow}{snow}

\item{snowReservoir}{snowReservoir}

\item{soilMoisture}{soilMoisture}

\item{soilWater}{soilWater}

\item{UH}{UH}

\item{soilDischarge}{soilDischarge}

\item{ddistAll}{ddistAll}

\item{groundwater}{groundwater}

\item{models}{models}

\item{pathResults}{path to save the results}

\item{saveDate}{date to be saved if any}
}
\description{
Simulation of the hydrological model
}
\examples{
\dontrun{
do.simTS()
}
}
\keyword{ddd}

