% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.snow.R
\name{do.snow}
\alias{do.snow}
\title{state of the information related to snow}
\usage{
do.snow(htempX, hprecipX, scaobX, snowX, modelSnow, modelPrecipLZ)
}
\arguments{
\item{htempX}{temperature for each elevation zone}

\item{hprecipX}{precipitation for each elevation zone}

\item{scaobX}{snow coverage observation}

\item{snowX}{snow}

\item{modelSnow}{list(nbLevelZone,unitsnow,n0,Ws,TS,CX,CFR,CGLAC,gca,UP)}

\item{modelPrecipLZ}{list(nbLevelZone,Plr,hfelt,midmetp)}
}
\description{
Update the information related to snow
}
\examples{
\dontrun{
do.snow()
}
}
\keyword{ddd}

