% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.snow.reservoir.R
\name{do.snow.reservoir}
\alias{do.snow.reservoir}
\title{state of Snow Reservoir}
\usage{
do.snow.reservoir(snow)
}
\arguments{
\item{snow}{snow}
}
\description{
Compute the snow reservoir
}
\examples{
\dontrun{
do.snow.reservoir()
}
}
\keyword{ddd}

