% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.stateX.R
\name{do.stateX}
\alias{do.stateX}
\title{State of the hydrological processes}
\usage{
do.stateX(Timeresinsec, precipX, tempX, scaobX, snowX, soilMoistureX,
  snowReservoirX, soilWaterX, UH, ddistAllX, soilDischargeX, groundwaterX,
  modelPrecipLZ, modelTempLZ, modelSnow, modelSoilMoisture, modelSoil, modelBog,
  modelET, modelSoilWater, modelArea, modelLayer)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{precipX}{precipitation value}

\item{tempX}{temperature value}

\item{scaobX}{observed snow coverage}

\item{snowX}{snow}

\item{soilMoistureX}{soilMoisture}

\item{snowReservoirX}{snowReservoir}

\item{soilWaterX}{soilWater}

\item{UH}{UH}

\item{ddistAllX}{ddistAll}

\item{soilDischargeX}{soilDischarge}

\item{groundwaterX}{groundwater}

\item{modelPrecipLZ}{list of parameters about PrecipLZ
list(nbLevelZone,Plr,hfelt,midmetp)}

\item{modelTempLZ}{list of parameters about TempLZ
list(nbLevelZone,Tlr,hfelt,midmett)}

\item{modelSnow}{list of parameters about Snow
list(nbLevelZone,unitsnow,n0,Ws,TS,CX,CFR,CGLAC,gca,UP)}

\item{modelSoilMoisture}{list of parameters about soilMoisture
list(gtcel,CapacityUpperLevel=2000,mLam,varLam,distr)}

\item{modelSoil}{list of parameters about Soil
list(glacfrac)}

\item{modelBog}{list of parameters about Bog
list(maxL,speed,nbStepsDelay,z,distr,param)}

\item{modelET}{list of parameters about Evapotranspiration
list(cea)}

\item{modelSoilWater}{list of parameters about Soil water
list(R)}

\item{modelArea}{list of parameters about Area
list(totarea,slopesriverarea,nobognoglacarea,bogarea)}

\item{modelLayer}{list of parameters about Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}
}
\description{
The function does the following:
 -state of Snow
- state of water volume in Soil Moisture: Z(t)
- state of Soil Water: EA(t), G(t), X(t) for soil and bogs
- state of Soil Discharge: D(t), runoff
- state of ddistAll: ddist, ddisx, S
- state of Groundwater: MAGKAP,M,Saturation Layers
}
\examples{
\dontrun{
do.stateX()
}
}
\keyword{ddd}

