% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_sim.R
\name{edd_sim}
\alias{edd_sim}
\title{Function to simulate the evolutionary distinctiveness dependent
diversification process}
\usage{
edd_sim(pars, age, model = "dsce2", metric = "ed", offset = "none")
}
\arguments{
\item{pars}{Vector of parameters: \cr \cr \code{pars[1]} corresponds to
lambda (speciation rate) \cr \code{pars[2]} corresponds to mu (extinction
rate) \cr \code{pars[3]} corresponds to beta_num (coefficient for species
number effect on speciation) \cr \code{pars[4]} corresponds to beta_phi
(coefficient for evolutionary distinctiveness effect on speciation)
\cr \code{pars[5]} corresponds to gamma_num (coefficient for species number
effect on speciation) \cr \code{pars[6]} corresponds to gamma_phi
(coefficient for evolutionary distinctiveness effect on extinction)}

\item{age}{Sets the crown age for the simulation}

\item{model}{Sets the model of diversity-dependence: \cr \code{model ==
dsce2} : linear dependence in speciation rate with parameters beta_num and
beta-phi\cr \code{model == dsde2} : linear dependence
in both speciation rate and extinction rate with parameters beta_num,
beta_phi, gamma_num and gamma_phi}

\item{metric}{Specifies which phylogenetic diversity metric should be used}

\item{offset}{Specifies which method to use to offset time effect on evolutionary
distinctiveness}
}
\value{
\item{ out }{ A list with the following four elements: The first
element is the tree of extant species in phylo format \cr The second element
is the tree of all species, including extinct species, in phylo format \cr
The third element is a matrix of all species where \cr - the first column is
the time at which a species is born \cr - the second column is the label of
the parent of the species; positive and negative values only indicate
whether the species belongs to the left or right crown lineage \cr - the
third column is the label of the daughter species itself; positive and
negative values only indicate whether the species belongs to the left or
right crown lineage \cr - the fourth column is the time of extinction of the
species. If this equals -1, then the species is still extant.\cr The fourth
element is the set of branching times of the tree of extant species.\cr The
fifth element is the lineage-through-time plot. \cr The sixth element is a
matrix of all evolutionary distinctiveness values of all lineages. \cr The
seventh element is a matrix of all the speciation and extinction rates of all
lineages. \cr The eighth element is a matrix of}
}
\description{
Simulating the evolutionary distinctiveness dependent
diversification process
}
\examples{
edd_sim(
  pars = c(0.5, 0.1, -0.001, -0.001, 0.001, 0.001), age = 6,
  model = "dsde2", metric = "ed", offset = "none"
)
}
\references{
- Etienne, R.S. et al. 2012, Proc. Roy. Soc. B 279: 1300-1309,
doi: 10.1098/rspb.2011.1439 \cr - Etienne, R.S. & B. Haegeman 2012. Am. Nat.
180: E75-E89, doi: 10.1086/667574
}
\author{
Tianjian Qin, Rampal S. Etienne
}
\keyword{models}
