% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edd_sim.R
\name{edd_simulation}
\alias{edd_simulation}
\title{Start an edd simulation}
\usage{
edd_simulation(combo = NULL, history = FALSE, verbose = FALSE, nrep = 5)
}
\arguments{
\item{combo}{A list containing parameter sets to be used to start a simulation}

\item{history}{Logical, decides whether to record historical states for all
of the lineages, including the transitions of the rates and the evolutionary
relationship metrics}

\item{verbose}{Logical, decides whether to print simulation details of each
time step during simulation, for debugging purposes}

\item{nrep}{Number of replications}
}
\value{
Returns a large list containing all the results
}
\description{
Function to start an edd simulation, of which the evolution
process is dependent on phylogenetic relationship among the community
}
\author{
Tianjian Qin
}
