% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{edd_simulation_plot}
\alias{edd_simulation_plot}
\title{Plotting results of a replicated edd simulation}
\usage{
edd_simulation_plot(
  result,
  pars,
  age,
  model,
  offset,
  EDPars,
  EDModel,
  EDMethod
)
}
\arguments{
\item{result}{a list of extracted results}

\item{pars}{the parameters used in simulation being plotted}

\item{age}{the total simulation time}

\item{model}{the model used in the simulation}

\item{EDPars}{the ED related parameters used in simulation being plotted}

\item{EDModel}{the ED model used in the simulation}

\item{EDMethod}{the method of ED being used in the simulation}
}
\value{
a plot grid generated by cow_plot
}
\description{
Function to automatically produce several plots from the results
of a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
