% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{edd_simulation_replicated}
\alias{edd_simulation_replicated}
\title{Running a replicated edd simulation}
\usage{
edd_simulation_replicated(rep, pars, age, model, metric, offset)
}
\arguments{
\item{rep}{the number of replication}

\item{pars}{parameters used in a edd simulation}

\item{age}{the total running time of a edd simulation}

\item{model}{the type of model used in a edd simulation}

\item{metric}{the metric used in a edd simulation}

\item{offset}{the offset method for Phi used in a edd simulation}
}
\value{
a list of simulation results
}
\description{
Function to automatically run a replicated edd simulation
}
\author{
Tianjian Qin
}
\keyword{distinctiveness}
\keyword{evolutionary}
\keyword{phylogenetics}
