% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{extract_edd_result}
\alias{extract_edd_result}
\title{Extracting all results generated in simulation}
\usage{
extract_edd_result(result, nrep, which, nlist = 8)
}
\arguments{
\item{result}{a list generated by edd simulation function}

\item{nrep}{the number of replication used in the simulation}
}
\value{
a list of extracted results
}
\description{
Function to extract all results generated in edd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
