% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{extract_pdd_result}
\alias{extract_pdd_result}
\title{Extracting all results generated in simulation}
\usage{
extract_pdd_result(result, nrep, which, nlist = 5)
}
\arguments{
\item{result}{a list generated by pdd simulation function}

\item{nrep}{the number of replication used in the simulation}
}
\value{
a list of extracted results
}
\description{
Function to extract all results generated in pdd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
