% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{get_Kprime}
\alias{get_Kprime}
\title{Get carrying capacity from other parameters of the model}
\usage{
get_Kprime(ddmodel, pars)
}
\arguments{
\item{ddmodel}{a character or integer specifying a DD model, as described in
dd_loglik() and dd_ML() documentation}

\item{pars}{a numeric vector containing parameter values of the DD model.
\code{pars[1]} is the lambda0, \code{pars[2]} is mu0, \code{pars[3]} is K
and \code{pars[4]}, if relevant, is r.}
}
\value{
a numeric value, K'
}
\description{
Compute the carrying capacity K', the maximum diversity that is possible to reach
with the model, i.e. the value of N for which lambda(N) = 0.
This is different from K, the equilibrium diversity, i.e. the value of N for 
which lambda(N) = mu(N).
}
\author{
Theo Pannetier
}
