% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Get the models to run the hydrological model}
\usage{
getModel(getModel, path, Timeresinsec = NULL, SAVE = FALSE,
  pathResults = "~")
}
\arguments{
\item{getModel}{method to get the model parameters ("build", "load" or "source")}

\item{path}{directory where to get the files}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{SAVE}{Boolean to save the current model}

\item{pathResults}{path to save the results}
}
\description{
Get all the models
Three options:
- build from a set of parameters
- load from a rda file
- load from a R-file
}
\examples{
\dontrun{
getModel()
}
}
\keyword{ddd}

