% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.ts.R
\name{graph.ts}
\alias{graph.ts}
\title{plot of the timeseries}
\usage{
graph.ts(dateTS, precip, q, q2 = NULL, minq = NULL, maxq = NULL,
  minprecip = NULL, maxprecip = NULL, name = "precip-discharge timeserie",
  path = NULL, width_mm = 180, height_mm = 180)
}
\arguments{
\item{dateTS}{POSIXct}

\item{precip}{precipitation timeserie}

\item{q}{runoff timeserie}

\item{q2}{runoff timeserie}

\item{minq}{minimum runoff value}

\item{maxq}{maximum runoff value}

\item{minprecip}{minimum precipitation value}

\item{maxprecip}{maximum precipitation values}

\item{name}{name of the chart}

\item{path}{path of the saved chart. If NULL, plot on screen}

\item{width_mm}{width of the chart}

\item{height_mm}{height of the chart}
}
\description{
Very simple timeserie chart
}
\examples{
\dontrun{
graph.ts()
}
}
\keyword{ddd}

