% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.groundwater.R
\name{init.groundWater}
\alias{init.groundWater}
\title{init
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers}
\usage{
init.groundWater(Timeresinsec, UHMAD, MAD, area, modelSaturation, modelLayer)
}
\arguments{
\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{area}{in squared meters}

\item{modelSaturation}{list of parameters about the saturation
list(gtcel,CapacityUpperLevel,mLam,varLam,distr)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}
}
\description{
init
The funtion initializes the main features of the groundwater:
- The maximum capacity of each staturation level,
- the Groundwater Storage Capacity (GSC),
- The saturation layers
}
\examples{
\dontrun{
init.groundWater()
}
}
\keyword{ddd}

