% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.initialConditions.R
\name{init.initialCondition}
\alias{init.initialCondition}
\title{ddd.init.initialCondition}
\usage{
init.initialCondition(initialCondition, path = NULL, MAD = NULL,
  q1 = NULL, Timeresinsec = NULL, nbLevelZone = NULL, NoL = NULL,
  D = NULL, UHriver = NULL, layerUH = NULL, UHMAD = NULL,
  modelArea = NULL, modelLayer = NULL, modelRiver = NULL,
  modelBog = NULL, modelSaturation = NULL)
}
\arguments{
\item{initialCondition}{method for the initialization, "load", "source", "default"}

\item{path}{path of the directory where to find the two files}

\item{MAD}{value of the Mean Annual Discharge}

\item{q1}{first value of the runoff timeserie}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{nbLevelZone}{number of level zone}

\item{NoL}{number of saturation layers}

\item{D}{volume of the unsaturated zone}

\item{UHriver}{Unit Hydrograp of the river}

\item{layerUH}{Unit Hydrograph of the saturation layers}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{modelArea}{list of parameters about the area
list(totarea,slopesriverarea,nobognoglacarea,bogarea)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}

\item{modelRiver}{list of parameters about the river
list(maxL,speed,nbStepsDelay,z,distr,param)}

\item{modelBog}{list of parameters of the bog
list(maxL,speed,nbStepsDelay,z,dist,param)}

\item{modelSaturation}{list of parameters of the saturation
list(gtcel,CapacityUpperLevel=2000 ,mLam,varLam,distr)}
}
\value{
The output is a list of the seven main blocks that will be process along the simulation:
 - snow
 - snowReservoir
 - soilMoisture
 - soilWater
 - soilDischarge
 - ddistAll
 - groundwater
}
\description{
The funtion \code{init.initialCondition()} initializes all the processess:
 i- snow (isoil,gisoil,spd,wcd,sca,nsno,alfa,ny,snowfree)
 ii- snow reservoir (snomag,swe_h,middelsca,snofritt)
 iii- soil moisture (waterSoil,waterGlacialSoil,waterGlaciers,Z)
 iv- soil water (G,X,Ea) (soil and bogs)
 v- soil discharge (D,qsimutx,qsimX)
 vi- ddistx, ddist, S
 vii- groundwater (Magkap, M, Layers)
}
\examples{
\dontrun{
init.initialCondition()
}
}
\keyword{ddd}

