% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize the hydrological processes}
\usage{
init(ddd, Timeresinsec, q1)
}
\arguments{
\item{ddd}{ddd object}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{q1}{initial runoff}
}
\description{
The function does the following:
 -initialize UH
- initialize snow
- initialize snowReservoir
- initialize soil moisture
- initialize soil discharge
- initialize ddistAll
- initialize grounwater
- initialize soil water
}
\examples{
\dontrun{
init()
}
}
\keyword{ddd}

