% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{L2ED}
\alias{L2ED}
\title{Converting a table with speciation and extinction events to evolutionary 
distances}
\usage{
L2ED(L, t)
}
\arguments{
\item{L}{Matrix of events as produced by pdd_sim: \cr \cr - the first column
is the time at which a species is born in Mya\cr - the second column is the
label of the parent of the species; positive and negative values indicate
whether the species belongs to the left or right crown lineage \cr - the
third column is the label of the daughter species itself; positive and
negative values indicate whether the species belongs to the left or right
crown lineage \cr - the fourth column is the time of extinction of the
species; if the fourth element equals -1, then the species is still extant.}

\item{t}{Simulation time}
}
\value{
a named vector of mean evolutionary distinctivenesses
}
\description{
Function to convert a table with speciation and extinction events to 
mean evolutionary distances between each species and the rest of the community
}
\author{
Tianjian Qin
}
\keyword{models}
