% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{L2Phi_cpp}
\alias{L2Phi_cpp}
\title{C++ version of L2Phi
#' @description Function to convert a table with speciation and extinction events to a
phylogenetic diversity metric}
\usage{
L2Phi_cpp(L, t, metric)
}
\arguments{
\item{L}{Matrix of events as produced by pdd_sim: \cr \cr - the first column
is the time at which a species is born in Mya\cr - the second column is the
label of the parent of the species; positive and negative values indicate
whether the species belongs to the left or right crown lineage \cr - the
third column is the label of the daughter species itself; positive and
negative values indicate whether the species belongs to the left or right
crown lineage \cr - the fourth column is the time of extinction of the
species; if the fourth element equals -1, then the species is still extant.}

\item{t}{Sets whether the phylogeny should drop species that are
extinct at the present}

\item{metric}{Specifies which phylogenetic diversity metric should be used}
}
\value{
a value of one of the phylogenetic diversity metrices
}
\description{
C++ version of L2Phi
#' @description Function to convert a table with speciation and extinction events to a
phylogenetic diversity metric
}
\author{
Tianjian Qin
}
\keyword{models}
