% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{ddd general simulation}
\usage{
main(fromPeriod = NULL, toPeriod = NULL, timeResolution = NULL,
  catchment = "catchment", pathData = NULL, fileData = NULL,
  pathParam = NULL, fileParam = NULL, methodParam = "processedNVE",
  methodModel = "processedNVE", pathResults = "~/", D_ci = 2,
  saveDate = NULL, FIGURE = FALSE)
}
\arguments{
\item{fromPeriod}{"YYYYmmddHH"}

\item{toPeriod}{"YYYYmmddHH"}

\item{timeResolution}{"daily", "three-hourly", "hourly"}

\item{catchment}{name of the catchement. Not in used yet}

\item{pathData}{path of the data file (ptq)}

\item{fileData}{name of the data file (ptq)}

\item{pathParam}{path of the parameters file}

\item{fileParam}{name of the parameters file (ptq)}

\item{methodParam}{method how to get the param. Default is "processedNVE"}

\item{methodModel}{method how to get the model. Default is "processedNVE"}

\item{pathResults}{path of the results. Default is HOME}

\item{D_ci}{initial condition for the potential volume of water that is needed for complete saturation. Default is 2}

\item{saveDate}{full data of this date will be saved}

\item{FIGURE}{Boolean to plot a basic plot. Default is FALSE}
}
\description{
Main function for the simulation of the hydrological model ddd
}
\examples{
\dontrun{
main()
}
}
\keyword{ddd}

