% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata8880.R
\docType{data}
\name{metadata8880}
\alias{metadata8880}
\title{Metadata of station no 8880}
\format{A csv file}
\source{
\href{http://www.met.no}{met.no}
}
\description{
Stations no 8880 is registered from the Norwegian Meteorlogical Institut
The metadata provides the information about:
 -the station id:Stnr
 -the station name: Name
 -when it starts to to be operational:Operates from
 - when it strops its operationability:Operates until
 - elevation: Altitude
 - Lattitude: Latitude
 - Longitude: Longitude
 - City: Municipality
 - County: County
}
\references{
the Norwegian Meteorological Institute (met.no)
}
\keyword{datasets}

