% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Build the models of the hydrological processes}
\usage{
model(ddd, method, ...)
}
\arguments{
\item{ddd}{ddd object}

\item{method}{"manual", "load","processedNVE"}

\item{...}{parameters related to the method}
}
\description{
Build the models of the hydrological processes
}
\examples{
\dontrun{
model()
}
}
\keyword{ddd}

