% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{obs}
\alias{obs}
\title{Get the observations}
\usage{
obs(ddd, pathPrecip, filenamePrecip, pathTemp, filenameTemp, pathQ, filenameQ)
}
\arguments{
\item{ddd}{ddd object}

\item{pathPrecip}{path of the precipitation observation}

\item{filenamePrecip}{name of the precipitation file}

\item{pathTemp}{path of the temperature observation}

\item{filenameTemp}{name of the temperature file}
}
\description{
Get the observations
}
\examples{
\dontrun{
obs()
}
}
\keyword{ddd}

