% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{param}
\alias{param}
\title{Get the input parameters of the hydrological processes}
\usage{
param(ddd, method, ...)
}
\arguments{
\item{ddd}{ddd object}

\item{method}{"manual", "load","processedNVE"}

\item{...}{parameters related to the method}
}
\description{
Get the input parameters of the hydrological processes
}
\examples{
\dontrun{
param()
}
}
\keyword{ddd}

