% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{pdd_simulation_plot}
\alias{pdd_simulation_plot}
\title{Plotting results of a replicated pdd simulation}
\usage{
pdd_simulation_plot(result, pars, age, model, offset)
}
\arguments{
\item{result}{a list of extracted results}

\item{pars}{the parameters used in simulation being plotted}

\item{age}{the total simulation time}

\item{model}{the model used in the simulation}

\item{offset}{the offset method for Phi used in a pdd simulation}
}
\value{
a plot grid generated by cow_plot
}
\description{
Function to automatically produce several plots from the results
of a replicated pdd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
