% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{pdd_simulation_replicated}
\alias{pdd_simulation_replicated}
\title{Running a replicated pdd simulation}
\usage{
pdd_simulation_replicated(rep, pars, age, model, metric, offset)
}
\arguments{
\item{rep}{the number of replication}

\item{pars}{parameters used in a pdd simulation}

\item{age}{the total running time of a pdd simulation}

\item{model}{the type of model used in a pdd simulation}

\item{metric}{the metric used in a pdd simulation}

\item{offset}{the offset method for Phi used in a pdd simulation}
}
\value{
a list of simulation results
}
\description{
Function to automatically run a replicated pdd simulation
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
