% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{phylo2mpd}
\alias{phylo2mpd}
\title{Converting a phylogenetic tree of phylo object to mean pairwise distance
among all tips}
\usage{
phylo2mpd(phy)
}
\arguments{
\item{phy}{a phylo object that only contains extant species
extinct at the present}
}
\value{
a numeric value of mean pairwise distance
}
\description{
Function to convert a phylogenetic tree with extant species to mean pairwise
distance
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
