% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{phylo2pd}
\alias{phylo2pd}
\title{Converting a phylogenetic tree of phylo object to total lengths of all
branches in the tree}
\usage{
phylo2pd(phy)
}
\arguments{
\item{a}{phylo object that only contains extant species
extinct at the present}
}
\value{
a numeric value of phylogenetic diversity
}
\description{
Function to convert a phylogenetic tree with extant species to
phylogenetic diversity
}
\author{
Tianjian Qin
}
\keyword{phylogenetics}
