% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{range01}
\alias{range01}
\title{standardizing data to the range from 0 to 1}
\usage{
range01(x)
}
\arguments{
\item{x}{Data to be standardized}
}
\description{
Function to standardize data to the range from 0 to 1 while retaining 
the relative size
}
\author{
Tianjian Qin
}
\keyword{standardize}
