% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{simplex}
\alias{simplex}
\title{Carries out optimization using a simplex algorithm (finding a minimum)}
\usage{
simplex(fun, trparsopt, optimpars, verbose = TRUE, ...)
}
\arguments{
\item{fun}{Function to be optimized}

\item{trparsopt}{Initial guess of the parameters to be optimized}

\item{optimpars}{Parameters of the optimization: relative tolerance in
function arguments, relative tolerance in function value, absolute tolerance
in function arguments, and maximum number of iterations}

\item{verbose}{adds verbose output of intermediate evaluations}

\item{...}{Any other arguments of the function to be optimimized, or settings
of the optimization routine}
}
\value{
\item{out}{ A list containing optimal function arguments
(\code{par}, the optimal function value (\code{fvalues}) and whether the
optimization converged (\code{conv})}.
}
\description{
Function to optimize target function using a
simplex method adopted from Matlab
}
\examples{

cat("No examples")

}
\author{
Rampal S. Etienne
}
\keyword{models}
