% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulation}
\usage{
simulation(ddd, timePeriod, saveDate, pathRes)
}
\arguments{
\item{ddd}{ddd object}

\item{timePeriod}{data timeserie and its related information}

\item{saveDate}{date to be saved if any}

\item{pathRes}{path of the results for the csv file}
}
\description{
Simulation of the hydrological model.
}
\examples{
\dontrun{
do.simTSv2()
}
}
\keyword{ddd}

