% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start}
\usage{
start(namefield, namespace, pathRes)
}
\arguments{
\item{namefield}{nme of the field}

\item{namespace}{namespace}

\item{pathRes}{path of the results}
}
\description{
Start by building the ddd Object
}
\examples{
start(namefield="ddd",namespace="ddd",pathRes="~/")

}
\keyword{ddd}

