% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.R
\name{stateX}
\alias{stateX}
\title{State of the hydrological processes}
\usage{
stateX(ddd, temperature, precipitation, scaob, Timeresinsec)
}
\arguments{
\item{ddd}{ddd object}

\item{temperature}{temperature input}

\item{precipitation}{precipitation input}

\item{scaob}{snow coverage input}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}
}
\description{
The function does the following:
 -state of Snow
- state of water volume in Soil Moisture: Z(t)
- state of Soil Water: EA(t), G(t), X(t) for soil and bogs
- state of Soil Discharge: D(t), runoff
- state of ddistAll: ddist, ddisx, S
- state of Groundwater: MAGKAP,M,Saturation Layers
}
\examples{
\dontrun{
stateX()
}
}
\keyword{ddd}

