% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station8880.R
\docType{data}
\name{station8880}
\alias{station8880}
\title{Station no 8880 with hourly precipitation and temperature}
\format{A csv file}
\source{
\href{http://met.no}{met.no}
}
\description{
Stations no 8880 is registered from the Norwegian Meteorlogical Institut
Hourly precipitation and temperature are made available between August 2 2016
and September 29 2016. No post-processing (qualicty control,...) has been done on the data.
}
\references{
the Norwegian Meteorological Institute (met.no)
}
\keyword{datasets}

