% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{transform_pars}
\alias{transform_pars}
\title{Transforming parameters from -Inf to Inf into parameters
from -1 to 1}
\usage{
transform_pars(pars)
}
\arguments{
\item{pars}{Parameters to be transformed}
}
\value{
Transformed parameters
}
\description{
Function to transform pars in a way that is more
useful for optimization: trpars <- sign(pars) * pars/(sign(pars) + pars);
}
\author{
Rampal S. Etienne
}
