% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{untransform_pars}
\alias{untransform_pars}
\title{Untransforming parameters from -1 to 1 into parameters
from -Inf to Inf.}
\usage{
untransform_pars(trpars)
}
\arguments{
\item{trpars}{Parameters to be untransformed}
}
\value{
Untransformed parameters
}
\description{
Function to untransform pars after optimization:
pars <- sign(trpars) * trpars/(sign(trpars) - trpars);
}
\author{
Rampal S. Etienne
}
