% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_24.9_24hptq_kal.R
\docType{data}
\name{val_24.9_24hptq_kal}
\alias{val_24.9_24hptq_kal}
\title{Dataset of daily variables ptq at Tingvatn}
\format{A txt file}
\source{
\href{http://www.met.no}{met.no}
}
\description{
Dataset of daily variables ptq at Tingvatn
}
\references{
the Norwegian Meteorological Institute (met.no)
}
\keyword{datasets}

