% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{what_is_this_ddmodel}
\alias{what_is_this_ddmodel}
\title{Helper function to quickly find what a ddmodel contains}
\usage{
what_is_this_ddmodel(ddmodel, short = FALSE)
}
\arguments{
\item{ddmodel}{a integer code between 1 and 15, corresponding to one of the
diversity-dependent models taken as input e.g. in [dd_ML()] and [dd_loglik()].}

\item{short}{logical. If FALSE (the default), returns a description of the
diversity-dependent functions used for speciation and extinction. If TRUE,
instead returns a 2-letter  code (speciation + extinction) summarising these 
functions: L for liner DD, P for a power DD function, X for an exponential DD
function, and C for a constant rate (no DD).}
}
\description{
Given a ddmodel code, returns a short description of the contents of the model
}
\author{
Theo Pannetier
}
