\name{DDD-internal}
\alias{dd_loglik_choosepar}
\alias{bd_loglik_choosepar}
\alias{dd_SR_loglik_choosepar}
\alias{dd_KI_loglik_choosepar}
\alias{dd_loglik_rhs}
\alias{dd_logliknorm_rhs1}
\alias{dd_logliknorm_rhs2}
\alias{dd_simplex}
\alias{bd_simplex}
\alias{dd_SR_simplex}
\alias{dd_KI_simplex}
\alias{L2phylo}
\title{Internal DDD Functions}
\description{Internal DDD functions}
\details{These are not to be called by the user.}
\keyword{internal}
