\name{DDD-internal}
\alias{dd_loglik_choosepar}
\alias{dd_SR_loglik_choosepar}
\alias{dd_KI_loglik_choosepar}
\alias{dd_loglik_rhs}
\alias{dd_logliknorm_rhs1}
\alias{dd_logliknorm_rhs2}
\alias{optimx2}
\title{Internal DDD Functions}
\description{Internal DDD functions}
\details{These are not to be called by the user.}
\keyword{internal}
