\name{DDD-internal}
\alias{initparsoptdefault}
\alias{parsfixdefault}
\alias{dd_loglik_high_lambda}
\alias{dd_loglik1}
\alias{dd_loglik2}
\alias{dd_loglik_choosepar}
\alias{bd_loglik_choosepar}
\alias{dd_SR_loglik1}
\alias{dd_SR_loglik2}
\alias{dd_SR_loglik_choosepar}
\alias{dd_KI_loglik1}
\alias{dd_KI_loglik2}
\alias{dd_KI_loglik_choosepar}
\alias{dd_MS_loglik1}
\alias{dd_MS_loglik2}
\alias{dd_MS_loglik_choosepar}
\alias{dd_loglik_rhs}
\alias{dd_loglik_M}
\alias{dd_loglik_M_bw}
\alias{dd_loglik_M_aux}
\alias{dd_loglik_M_bw_aux}
\alias{dd_loglik_M2}
\alias{dd_loglik_M2_aux}
\alias{dd_loglik_M3}
\alias{dd_loglik_M3_aux}
\alias{dd_KI_loglikold}
\alias{dd_loglik_bw_rhs}
\alias{dd_logliknorm_rhs1}
\alias{dd_logliknorm_rhs2}
\alias{dd_logliknorm_rhs3}
\alias{td_loglik_rhs}
\alias{td_loglik_bw_rhs}
\alias{td_loglik_rhs_sim}
\alias{td_sim}
\alias{dd_simplex}
\alias{bd_simplex}
\alias{dd_SR_simplex}
\alias{dd_KI_simplex}
\alias{dd_MS_simplex}
\alias{brts2phylo}
\alias{conv}
\alias{flavec}
\alias{lambdamu}
\alias{lambdamu2}
\alias{lambdamu3}
\alias{expn_dd_sp}
\alias{predictsp}
\alias{expn_dd}
\alias{lamu_dd_td}
\alias{dd_lamuN}
\alias{dd_MS_lamuN}
\alias{dd_KI_lamuN}
\alias{dd_bootstrap_sim}
\alias{simplex}
\alias{lambdamu}
\alias{changepars}
\alias{transform_pars}
\alias{untransform_pars}
\alias{expfun}
\alias{logH}
\alias{logLL}
\alias{dd_loglik_test}
\alias{dd_integrate}
\alias{dd_ode_FORTRAN}
\alias{dd_loglik_rhs_precomp}
\alias{dd_fill1d}
\alias{dd_initmod}
\alias{dd_runmod}
\alias{dd_runmodbw}
\alias{dd_runmodtd}
\title{Internal DDD Functions}
\description{Internal DDD functions}
\details{These are not to be called by the user.}
\keyword{internal}
