% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDD-package.R
\docType{package}
\name{DDD-package}
\alias{DDD}
\alias{DDD-package}
\title{DDD: Diversity-Dependent Diversification}
\description{
Implements maximum likelihood and bootstrap methods based on the diversity-dependent birth-death process to test whether speciation or extinction are diversity-dependent, under various models including various types of key innovations. See Etienne et al. 2012, Proc. Roy. Soc. B 279: 1300-1309, \doi{10.1098/rspb.2011.1439}, Etienne & Haegeman 2012, Am. Nat. 180: E75-E89, \doi{10.1086/667574}, Etienne et al. 2016. Meth. Ecol. Evol. 7: 1092-1099, \doi{10.1111/2041-210X.12565} and Laudanno et al. 2021. Syst. Biol. 70: 389–407, \doi{10.1093/sysbio/syaa048}. Also contains functions to simulate the diversity-dependent process.
}
\author{
\strong{Maintainer}: Rampal S. Etienne \email{r.s.etienne@rug.nl} (\href{https://orcid.org/0000-0003-2142-7612}{ORCID})

Authors:
\itemize{
  \item Bart Haegeman \email{bart.haegeman@sete.cnrs.fr} (\href{https://orcid.org/0000-0003-2325-4727}{ORCID})
}

Other contributors:
\itemize{
  \item Hanno Hildenbrandt \email{h.hildenbrandt@rug.nl} (\href{https://orcid.org/0000-0002-6784-1037}{ORCID}) [contributor]
  \item Giovanni Laudanno \email{glaudanno@gmail.com} (\href{https://orcid.org/0000-0002-2952-3345}{ORCID}) [contributor]
}

}
\keyword{internal}
