% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.groundwater.R
\name{init.groundWater}
\alias{init.groundWater}
\title{init
The funtion initializes the main features of the groundwater:}
\usage{
init.groundWater(method = NULL, path = NULL, Timeresinsec = NULL,
  UHMAD = NULL, MAD = NULL, area = NULL, modelSaturation = NULL,
  modelLayer = NULL)
}
\arguments{
\item{method}{method for the initialization, "load", "source", "manual", "processed"}

\item{path}{directory where to get the files, in used when method is "load" or "source"}

\item{Timeresinsec}{time resolution of the process in second (1hour: 3600s, ... etc)}

\item{UHMAD}{Unit Hydrograph of the Mean Annual Discharge}

\item{MAD}{value of the Mean Annual Discharge}

\item{area}{in squared meters}

\item{modelSaturation}{list of parameters about the saturation
list(gtcel,CapacityUpperLevel,mLam,varLam,distr)}

\item{modelLayer}{list of parameters about the Layers
list(maxL,speed,nbStepsDelay,z,distr,param,NoL)}
}
\description{
init
The funtion initializes the main features of the groundwater:
}
\examples{
\dontrun{
init.groundWater()
}
}
\keyword{ddd}

