% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_utils.R
\name{L2ED}
\alias{L2ED}
\title{Converting a table with speciation and extinction events to evolutionary 
distances}
\usage{
L2ED(L, t)
}
\arguments{
\item{L}{Matrix of events as produced by pdd_sim: \cr \cr - the first column
is the time at which a species is born in Mya\cr - the second column is the
label of the parent of the species; positive and negative values indicate
whether the species belongs to the left or right crown lineage \cr - the
third column is the label of the daughter species itself; positive and
negative values indicate whether the species belongs to the left or right
crown lineage \cr - the fourth column is the time of extinction of the
species; if the fourth element equals -1, then the species is still extant.}
}
\value{
a named vector of evolutionary distances
}
\description{
Function to convert a table with speciation and extinction events to 
mean evolutionary distances between each species and the rest of the community
}
\references{
- Etienne, R.S. et al. 2012, Proc. Roy. Soc. B 279: 1300-1309,
doi: 10.1098/rspb.2011.1439 \cr - Etienne, R.S. & B. Haegeman 2012. Am. Nat.
180: E75-E89, doi: 10.1086/667574
}
\author{
Tianjian Qin
}
\keyword{models}
